Move SendStatus and related code to core.sendstatus
