+++ app/src/main/java/app/pachli/components/compose/ComposeViewModel.kt
@@ -27,7 +27,6 @@ import androidx.core.net.toUri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import app.pachli.R
-import app.pachli.components.compose.ComposeActivity.QueuedMedia
 import app.pachli.components.compose.ComposeAutoCompleteAdapter.AutocompleteResult
 import app.pachli.components.compose.UiError.SaveAttachmentError
 import app.pachli.components.compose.UploadState.Uploaded
@@ -54,17 +53,21 @@ import app.pachli.core.model.Status
 import app.pachli.core.navigation.ComposeActivityIntent.ComposeOptions
 import app.pachli.core.navigation.ComposeActivityIntent.ComposeOptions.ComposeKind
 import app.pachli.core.navigation.ComposeActivityIntent.ComposeOptions.ReferencingStatus
+import app.pachli.core.network.extensions.SaveUriError
+import app.pachli.core.network.extensions.isInDirectory
+import app.pachli.core.network.extensions.saveToDirectory
 import app.pachli.core.network.retrofit.MastodonApi
 import app.pachli.core.network.retrofit.apiresult.ApiError
 import app.pachli.core.preferences.SharedPreferencesRepository
 import app.pachli.core.preferences.ShowSelfUsername
+import app.pachli.core.sendstatus.MediaUploader
+import app.pachli.core.sendstatus.MediaUploaderError
+import app.pachli.core.sendstatus.QueuedMedia
+import app.pachli.core.sendstatus.UploadState
 import app.pachli.core.ui.MentionSpan
 import app.pachli.service.MediaToSend
 import app.pachli.service.ServiceClient
 import app.pachli.service.StatusToSend
-import app.pachli.util.SaveUriError
-import app.pachli.util.isInDirectory
-import app.pachli.util.saveToDirectory
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
 import com.github.michaelbull.result.Result
