+++ app/src/main/java/app/pachli/receiver/SendStatusBroadcastReceiver.kt
@@ -40,8 +40,8 @@ import app.pachli.core.data.repository.AccountManager
 import app.pachli.core.data.repository.DraftsRepository
 import app.pachli.core.designsystem.R as DR
 import app.pachli.core.model.Draft
-import app.pachli.service.SendStatusService
-import app.pachli.service.StatusToSend
+import app.pachli.core.sendstatus.SendStatusService
+import app.pachli.core.sendstatus.StatusToSend
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineScope
