+++ settings.gradle.kts
@@ -62,6 +62,7 @@ include(":core:preferences")
 include(":core:navigation")
 include(":core:network")
 include(":core:network-test")
+include(":core:sendstatus")
 include(":core:testing")
 include(":core:ui")
 include(":core:worker")
